refactor(orchestrator): extract consolidation-run coordinator (seam 17, #1526)
Extract the full consolidation maintenance pass into
ConsolidationRunCoordinator (orchestration/consolidation-run.ts).

Moved:
- runConsolidation() — 504-LOC main consolidation cycle (LLM
  merge/invalidate/update, entity merge, commitment/TTL cleanup,
  lifecycle policy, compression guideline learning, tier migration,
  fact archival, semantic consolidation, identity consolidation,
  profile consolidation, summarization, topic extraction, TMT rebuild)
- recordScheduledDreamsPhaseRun() — dream-phase ledger recording
  (static import replaces dynamic import)

Inlined into the coordinator body (single call-site thin delegates
to lifecyclePolicyCoordinator/compressionGuidelineCoordinator).

Pattern: deps-bundle + lazy-getter + injectable gates, matching the
extraction-run coordinator (seam 15). The orchestrator keeps a thin
delegating runConsolidation() so call sites (runConsolidationNow,
maintenance scheduler) and tests work unchanged.

Behavior-preserving move. No logic changes.

orchestrator.ts: 17425 -> 16925 LOC (-500)
directStorageImports: 42 (unchanged)
Ratchet baseline updated.

Refs #1526
